backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
faster-whisper
requests
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
huggingface-hub<0.14
pyannote.audio
soundfile
pydub
fastapi
uvicorn
python-multipart
nvidia-cublas-cu12
nvidia-cudnn-cu12